@@ -15,45 +15,45 @@
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "7.1.0",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.1.0",
-    "@angular/compiler": "7.1.0",
-    "@angular/core": "7.1.0",
-    "@angular/forms": "7.1.0",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.1.0",
-    "@angular/platform-browser-dynamic": "7.1.0",
-    "core-js": "^2.6.5",
-    "rxjs": "^6.4.0",
-    "zone.js": "0.8.26"
+    "@angular/animations": "8.2.6",
+    "@angular/cdk": "^8.2.0",
+    "@angular/common": "8.2.6",
+    "@angular/compiler": "8.2.6",
+    "@angular/core": "8.2.6",
+    "@angular/forms": "8.2.6",
+    "@angular/material": "^8.2.0",
+    "@angular/platform-browser": "8.2.6",
+    "@angular/platform-browser-dynamic": "8.2.6",
+    "core-js": "^2.6.9",
+    "rxjs": "^6.5.3",
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.8.9",
+    "@angular-devkit/build-angular": "^0.803.4",
     "@angular-devkit/schematics": "^7.3.8",
-    "@angular/cli": "7.1.0-rc.0",
-    "@angular/compiler-cli": "7.1.0",
-    "@angular/language-service": "7.1.0",
+    "@angular/cli": "8.3.4",
+    "@angular/compiler-cli": "8.2.6",
+    "@angular/language-service": "8.2.6",
     "@types/jasmine": "~3.0.0",
-    "@types/node": "^10.14.4",
-    "codelyzer": "~4.5.0",
+    "@types/node": "^10.14.18",
+    "codelyzer": "^5.1.0",
     "cpr": "^3.0.0",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.0",
     "karma": "^3.1.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.0",
     "karma-jasmine-html-reporter": "^1.0.0",
     "moment": "^2.24.0",
-    "ng-packagr": "4.4.0",
+    "ng-packagr": "^5.5.0",
     "protractor": "^5.4.2",
-    "rimraf": "^2.6.3",
+    "rimraf": "^2.7.1",
     "ts-node": "~7.0.0",
-    "tsickle": "^0.33.0",
-    "tslib": "^1.7.1",
+    "tsickle": "0.37.0",
+    "tslib": "^1.9.0",
     "tslint": "~5.11.0",
-    "typescript": "3.1.6"
+    "typescript": "3.5.3"
   }
 }
